Updates some Helm Chart aspects in order to ease the deployment process
- Added few default configurations to the values.yaml related to the
security context. This should make the configuration and deployment of
the resources into the RNDP environment easier.

- Added option to run the container with custom `command` and `args` in
order to allow troubleshooting.

- Removed couple of default agent configurations. They are not needed as
defaults and also in most cases are different, depending on the set up
and environment. They need to be provided externally in order to merge
them with the default configuration.