247 Introduces an ability to provide different agent configuration
- Added logic that allows the users to provide different agent
configuration based on a ConfigMap object. It will override the default
agent configuration, when set.

  In specific cases the default agent configuration + specific
properties overrides does not work. So we need to have an easy way to
set completely different agent configuration for specific environment.
This changes will allow us to do that without need of any workarounds.